resturctured instruction
